login_registration_added

+++ app/Http/Controllers/ReviewController.php
@@ -3,12 +3,16 @@
 namespace App\Http\Controllers;
 
 use App\Models\Review;
+use App\Models\User;
 use App\Models\Product;
 use App\Http\Requests\StoreReviewRequest;
 use App\Http\Requests\ReviewRequest;
 use App\Http\Requests\UpdateReviewRequest;
 use App\Http\Resources\ReviewResource;
 use Symfony\Component\HttpFoundation\Response;
+use App\Exceptions\UserNotLoggedIn;
+use App\Exceptions\ReviewNotBelongsToUser;
+use Auth;
 
 class ReviewController extends Controller
 {
@@ -41,13 +45,23 @@ class ReviewController extends Controller
      */
     public function store(ReviewRequest $request, Product $product)
     {
-        $review = new Review($request->all());
+        // $this->reviewUserCheck();
+       
+        $review = new Review();
+       
+        $review->star = $request->star;
+        $review->review = $request->review;
+        $review->product_id = $product->id;
+        $review->user_id = Auth::id();
         $product->reviews()->save($review);
 
         return response([
             'data'=>new ReviewResource($review)
         ], Response::HTTP_CREATED);
-    }
+    
+
+    
+}
 
     /**
      * Display the specified resource.
@@ -80,6 +94,9 @@ class ReviewController extends Controller
      */
     public function update(ReviewRequest $request,Product $product, Review $review)
     {
+
+        $this->checkUser($review);
+
         $review->update($request->all());
         
         return response([
@@ -95,10 +112,41 @@ class ReviewController extends Controller
      */
     public function destroy(Product $product, Review $review)
     {
+
+        if(auth()->check()) {
+        // if the review doesn't belong to user...
+        if(auth()->user()->id!=$review->user_id)
+        {
+            throw new ReviewNotBelongsToUser;
+        }
         $review->delete();
+       
 
         return response([
             null
         ], Response::HTTP_NO_CONTENT);
     }
+    throw new UserNotLoggedIn;
+}
+
+    public function reviewUserCheck() {
+        $users = User::all();
+        foreach($users as $user)
+        {
+            if(Auth::id() == $user->id)
+            return true;
+        }
+
+        throw new UserNotLoggedIn;
+    }
+
+
+    public function checkUser(Review $review){
+        if($review->user_id !=Auth::id())
+        {
+            throw new ReviewNotBelongsToUser;
+        }
+
+    }
+   
 }
